[WEB] Modified delete function
